add container (docker) only support for kruize
