+++ src/main/java/com/kruize/environment/docker/DockerEnvImpl.java
@@ -59,6 +59,19 @@ public class DockerEnvImpl extends EnvTypeImpl
     public void getAllApps()
     {
         JsonArray containerList = null;
+        /**
+         * kruize-docker.json contains the details of the containers.
+         * For each container, the details are expressed in the form of key-value pairs
+         * where the keys are name, cpu_limit and mem_limit.
+         * Example file:
+         * {
+         *   "containers": [
+         *     { "name": "kruize", "cpu_limit": "0.5", "mem_limit": "70m" },
+         *     { "name": "acmeair-mono", "cpu_limit": "3.3", "mem_limit": "350m" },
+         *     { "name": "acmeair-db1", "cpu_limit": "5", "mem_limit": "2.3g" }
+         *   ]
+         * }
+         */
         try (FileReader reader = new FileReader("/opt/app/kruize-docker.json"))
         {
             containerList = new JsonParser()
@@ -71,9 +84,9 @@ public class DockerEnvImpl extends EnvTypeImpl
             e.printStackTrace();
         }
 
-        if(containerList != null && containerList.size() > 0) {
-            for(JsonElement container : containerList) {
-                if(container != null && container.getAsJsonObject().size() > 0)
+        if (containerList != null && containerList.size() > 0) {
+            for (JsonElement container : containerList) {
+                if (container != null && container.getAsJsonObject().size() > 0)
                     insertContainerMetrics(container);
             }
         }
@@ -107,14 +120,11 @@ public class DockerEnvImpl extends EnvTypeImpl
         containerMetrics.setStatus("Running");
         containerMetrics.setApplicationName(containerName);
 
-        if (container.getAsJsonObject().has("mem_limit"))
-        {
+        if (container.getAsJsonObject().has("mem_limit")) {
             containerMetrics.setOriginalMemoryLimit(container.getAsJsonObject().get("mem_limit").getAsDouble());
         }
 
-        if (container.getAsJsonObject().has("cpu_limit"))
-        if (container.getAsJsonObject().has("cpu_limit"))
-        {
+        if (container.getAsJsonObject().has("cpu_limit")) {
             containerMetrics.setOriginalMemoryLimit(container.getAsJsonObject().get("cpu_limit").getAsDouble());
         }
